finished majority of project. Moved code to separate files and refactored code
