+++ src/App/css/default-popup.css
@@ -22,14 +22,21 @@ section {
 #home-container {
   background-color: rgb(255, 255, 255) !important;
   margin-top: 100px;
+  height: 450px;
 }
 
 #visualiser-container{
-  margin-top: 90px;
+  margin-top: 95px;
+  padding-bottom: 10px;
 }
 
 #countries-container{
   /* overflow: auto; */
+  margin-top: 5px;
+}
+
+#installation-container {
+  margin-top: 5px;
 }
 
 #Total-installs-container {
@@ -119,7 +126,7 @@ section {
 }
 .line1 {
   transform: rotate(330deg);
-  top: 44.2%;
+  top: 43.8%;
   left: 35%;
   background-color: #0b9e58;
 }
@@ -171,7 +178,7 @@ section {
     transform: rotate(180deg);
     left: 25%;
     border-radius: 0px;
-    top: 63.9%;
+    top: 64.1%;
   }
 }
 .line4 {
@@ -189,7 +196,7 @@ section {
     transform: rotate(180deg);
     right: 25%;
     border-radius: 0px;
-    top: 44.2%;
+    top: 43.8%;
   }
 }
 .line5 {
@@ -207,7 +214,7 @@ section {
     transform: rotate(180deg);
     right: 25%;
     border-radius: 0px;
-    top: 63.8%;
+    top: 64.2%;
     width: 150px;
   }
 }
@@ -257,6 +264,7 @@ section {
 
 #clear-data-button{
   margin-left: 380px;
+  margin-top: -50px;
 }
 
 #submit-countries {
@@ -320,24 +328,50 @@ section {
 
 }
 
-/* ::-webkit-scrollbar {
-  width: 18px;
-}
-
-::-webkit-scrollbar-thumb {
-  background-color: #9eb2b9;
-  border-radius: 20px;
-  border: 6px solid transparent;
-  background-clip: content-box;
-}
-
-::-webkit-scrollbar-thumb:hover {
-  background-color: #ccd9dd;
-} */
-
 .alert {
   position: fixed !important;
   bottom: 0;
   width: 100%;
   margin-bottom: 0 !important;
+}
+
+/* Countries placeholder style */
+#chartDivCountries {
+  width: 100%;
+  height: 500px;
+  margin: 0 auto;
+}
+
+#placeholder-countries {
+  width: 72%;
+  height: 400px;
+  margin: 0 auto;
+  display: block;
+}
+
+#placeholder-graph-spacing {
+  width: 50%;
+  height: 60px;
+  margin: 0 auto;
+}
+
+/* Total installations placeholder style */
+#chartDivInstallations {
+  width: 100%;
+  height: 500px;
+  margin: 0 auto;
+}
+
+#placeholder-installs {
+  width: 100%;
+  height: 400px;
+  margin: 0 auto;
+  display: block;
+}
+
+#chartDivInstallationsQuarters {
+  width: 100%;
+  height: 500px;
+  margin: 0 auto;
+  display: none;
 }
