--- src/App/javascript/charts/visualiser.js
@@ -1,4 +1,0 @@
-
-
-
-
