+++ src/App/javascript/helper/helper-functions.js
@@ -12,8 +12,26 @@ export function alertTimeout(id) {
   */
   return setTimeout(() => {
     document.getElementById(id).style.opacity = 0;
+    document.getElementById(id).classList.remove("alert-success");
     document.getElementById(id).classList.remove("alert-danger");
     document.getElementById(id).classList.remove("alert-warning");
-    document.getElementById(id).classList.remove("alert-success");
   }, 5000);
+}
+
+export function generateRandomLightColours(numberOfColours) {
+  function getRandomColor() {
+    let letters = "BCDEF".split("");
+    let color = "#";
+    for (let i = 1; i <= 6; i++) {
+      color += letters[Math.floor(Math.random() * letters.length)];
+    }
+    return color;
+  }
+
+  let coloursArray = [];
+  for (let j = 0; j < numberOfColours; j++) {
+    coloursArray.push(getRandomColor());
+  }
+
+  return coloursArray;
 }
