+++ src/App/pages/default-popup.html
@@ -2,7 +2,6 @@
 <html>
   <head>
     <title>Chrome extension stats</title>
-    <meta charset="utf-8" name="viewport" content="width=device-width, initial-scale=1"/>
     <!-- Our own style sheet -->
     <link href="/App/css/default-popup.css" rel="stylesheet" />
     <!-- BootStrap style sheet -->
@@ -10,7 +9,6 @@
       href="https://cdn.jsdelivr.net/npm/bootstrap@5.1.3/dist/css/bootstrap.min.css"
       rel="stylesheet"
     />
-
     <!-- External fonts -->
   </head>
 
@@ -89,14 +87,11 @@
       <div class="line line6"></div>
       <button type="button" id="clear-data-button" style="display: none;" class="btn btn-secondary">Clear data</button>
     </div>
-
-
-    <div style="width: 50%; height: 64px; margin: 0 auto"></div>
   </div>
+
   <!-- Visualiser container -->
-  <div class="custom-navbar">
+  <div>
     <div class="container-fluid" id="visualiser-container" style="display: none">
-      <div id="chartDiv" style="width: 50%; height: 10px; margin: 0 auto"></div>
       <!-- Sub navbar-->
       <nav class="nav nav-tabs nav-fill bg-light">
         <a
@@ -111,15 +106,9 @@
           >Total Installations</a
         >
       </nav>
-      <!-- Visualiser sub navbar Countries-->
+      <!-- Visualiser Countries tab -->
       <div class="container-fluid" id="countries-container" style="display: none">
-        <div style="width: 50%; height: 10px; margin: 0 auto"></div>
         <div class="d-grid gap-2">
-          <!--
-          <button class="btn btn-outline-primary" id="csv-countries-button-click" type="button" >
-            Upload csv All Countries
-          </button>-->
-
           <div class="input-group">
             <input
               type="file"
@@ -140,17 +129,15 @@
           <div
             class="bg-light border"
             id="chartDivCountries"
-            style="width: 100%; height: 500px; margin: 0 auto"
           >
             <div
               id="placeholder-countries"
-              style="width: 72%; height: 400px; margin: 0 auto; display: block"
             >
               <p class="lead" id="placeholder-countries-text">
                 No data for current period
               </p>
 
-              <div style="width: 50%; height: 60px; margin: 0 auto"></div>
+              <div id="placeholder-graph-spacing"></div>
               <canvas
                 id="placeholder-countries-image"
                 width="100%"
@@ -168,19 +155,13 @@
 
         </div>
       </div>
-      <!-- Visualiser sub navbar Installations-->
+      <!-- Visualiser Total Installations tab-->
       <div
         class="container-fluid"
         id="installation-container"
         style="display: none"
       >
-        <div style="width: 50%; height: 10px; margin: 0 auto"></div>
         <div class="d-grid gap-2">
-          <!--
-        <button class="btn btn-outline-primary" id="csv-countries-button-click" type="button" >
-          Upload csv All Countries
-        </button>-->
-
           <div class="input-group">
             <input
               type="file"
@@ -201,11 +182,9 @@
           <div
             class="bg-light border"
             id="chartDivInstallations"
-            style="width: 100%; height: 500px; margin: 0 auto"
           >
             <div
               id="placeholder-installs"
-              style="width: 100%; height: 400px; margin: 0 auto; display: block"
             >
               <p class="lead" id="placeholder-installs-text">
                 No data for current period
@@ -226,7 +205,6 @@
           <div
             class="bg-light border"
             id="chartDivInstallationsQuarters"
-            style="width: 100%; height: 500px; margin: 0 auto; display: none"
           >
           <canvas
               id="myChartInstallationsQuarters"
@@ -235,10 +213,10 @@
             ></canvas>
           </div>
         </div>
-
       </div>
     </div>
   </div>
+
   <!-- Alerts -->
   <div class="alert" id="alert" role="alert" style="opacity: 0"></div>
   <!-- JavaScript
@@ -247,8 +225,9 @@
     to getElementById for example, as that element will not exist yet.
   -->
   <script type="module" src="/App/javascript/default-popup.js"></script>
-  <script type="module" src="/App/javascript/csv-parse.js"></script>
-  <script type="module" src="/App/javascript/charts/visualiser.js"></script>
+  <script type="module" src="/App/javascript/csv-parse-home.js"></script>
+  <script type="module" src="/App/javascript/visualiser-pages/visualiser-countries-page.js"></script>
+  <script type="module" src="/App/javascript/visualiser-pages/visualiser-installations-page.js"></script>
   <!-- Chart js -->
   <script src="../../node_modules/chart.js/dist/chart.js"></script>
 </html>
